Moves NewCharacter to root Home folder (out of components
